project-evaluator/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.1
openai==1.82.1
python-dotenv==1.0.0
requests==2.31.0

# Multi-provider AI dependencies
anthropic>=0.52.2
google-genai>=1.21.1
databricks-sdk>=0.57.0
PyYAML>=6.0

# Azure OpenAI (uses standard openai library with different base URL)
# No additional package needed
